backend/pyproject.toml
[tool.poetry]
name = "dcf-teacher-afrotech"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
transformers = "^4.44.2"
fastapi = "^0.114.0"
uvicorn = "^0.30.6"
pydantic = "^2.9.0"
requests = "^2.32.3"
google-generativeai = "^0.7.2"
beautifulsoup4 = "^4.12.3"
python-dotenv = "^1.0.1"
pandas = "^2.2.2"


[[tool.poetry.source]]
name = "pytorch_cpu"
url = "https://download.pytorch.org/whl/cpu"
priority = "explicit"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
